requirements.txt
Pillow
pyautogui
pynput
pyinstaller
requests